Release/0.1.1 (#10)
This pull request adds
- Adds a new Reviews page
- Improves the existing product page, and adds a sorting feature for products. 
- Dockerfiles for the different components of the application 
- Updates the docker-compose.yml file to include the necessary services. 
- Updates scrapper to bring more data to the database
- Adds simple clean data pipe for `price` field

The Dockerfiles are used to build Docker images for each component, allowing for easier deployment and scalability. The docker-compose.yml file is updated to include the new services and their dependencies. 

This change improves the development and deployment process by providing a standardized and reproducible environment for running the application.

Also, the changes include updating the number of rows displayed on the product page, updating the author's information in the README file, updating the login functionality on the main page, and updating the sidebar links.